requirements.txt
aiogram>=3.0
openai>=1.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
python-dotenv>=1.0
